refactor(search): split state from rendering
